backend/requirements.txt
Django
djangorestframework
django-rest-knox
Pillow